Minor refactor: extract the control-panel component
