+++ widgets/htdocs/alphafold/proteinFeaturesControlPanel.js
@@ -1,22 +1,11 @@
 import { html, css, LitElement } from 'https://unpkg.com/lit@2.0.0-rc.4/index.js?module';
 
+import './controlPanel.js';
+
 export class ProteinFeaturesControlPanel extends LitElement {
 
   static get styles() {
     return css`
-      :host {
-        display: block;
-        border: 1px solid #ccc;
-        min-width: max-content;
-      }
-
-      .panel-title {
-        background-color: var(--main-dark);
-        color: white;
-        font-weight: bold;
-        padding: 4px;
-      }
-
       .section-summary {
         display: grid;
         grid-template-columns: 1fr auto auto auto;
@@ -164,10 +153,9 @@ export class ProteinFeaturesControlPanel extends LitElement {
 
   render() {
     return html`
-      <div class="panel-title">
-        Protein features
-      </div>
-      ${ this.renderSections() }
+      <control-panel title="Protein features">
+        ${ this.renderSections() }
+      </control-panel>
     `;
   }
